breaking change: all pages must be created with a decorator
