+++ nicegui/routes.py
@@ -56,22 +56,3 @@ def get(self, path: str):
         self.add_route(routing.Route(path, decorated))
         return decorated
     return decorator
-
-
-def private_page(self, path: str):
-    '''
-    Use as a decorator for a function like this:
-
-    @ui.private_page('/private')
-    def create_private_page():
-        ui.label(f'your private page {uuid4()}')
-    '''
-    def decorator(func):
-        @wraps(func)
-        async def decorated():
-            with Page(None) as page:
-                await func() if is_coroutine(func) else func()
-            return page
-        jp.Route(path, decorated)
-        return decorated
-    return decorator
